+++ example/table/align.dart
@@ -3,7 +3,7 @@ import 'package:text_table/text_table.dart';
 main() {
   {
     final Table tab = table(['Player', 'Team', 'Goals'],
-        columnWidths: [fixed(25), fixed(25), fixed(10)])
+        colWidths: [fixed(25), fixed(25), fixed(10)])
       ..row(['Messi', 'Barcelona FC', 80])
       ..row(['Christiano Ronaldo', 'Real Madrid', 30])
       ..row(['Luiz Suarez', 'Barcelona FC', 50]);
@@ -12,7 +12,7 @@ main() {
 
   {
     final Table tab = table(['Player', 'Team', 'Goals'],
-        columnWidths: [fixed(25), fixed(25), fixed(10)], align: Align.center)
+        colWidths: [fixed(25), fixed(25), fixed(10)], globalAlign: Align.center)
       ..row(['Messi', 'Barcelona FC', 80])
       ..row(['Christiano Ronaldo', 'Real Madrid', 30])
       ..row(['Luiz Suarez', 'Barcelona FC', 50]);
@@ -21,7 +21,7 @@ main() {
 
   {
     final Table tab = table(['Player', 'Team', 'Goals'],
-        columnWidths: [fixed(25), fixed(25), fixed(10)], align: Align.right)
+        colWidths: [fixed(25), fixed(25), fixed(10)], globalAlign: Align.right)
       ..row(['Messi', 'Barcelona FC', 80])
       ..row(['Christiano Ronaldo', 'Real Madrid', 30])
       ..row(['Luiz Suarez', 'Barcelona FC', 50]);
@@ -30,7 +30,7 @@ main() {
 
   {
     final Table tab = table(['Player', 'Team', 'Goals'],
-        columnWidths: [fixed(25), fixed(25), fixed(10)],
+        colWidths: [fixed(25), fixed(25), fixed(10)],
         aligns: [Align.left, Align.center, Align.right])
       ..row(['Messi', 'Barcelona FC', 80])
       ..row(['Christiano Ronaldo', 'Real Madrid', 30])