+++ example/table/ellipses.dart
@@ -1,0 +1,11 @@
+import 'package:text_table/text_table.dart';
+
+main() {
+  final Table tab = table(['Player', 'Team', 'Goals'])
+    ..row(['Messi', 'Barcelona FC', 80])
+    ..row(['Christiano Ronaldo', 'Real Madrid', 30])
+    ..row(['Luiz Suarez', 'Barcelona FC', 50]);
+  tab.maxColWidths[0] = fixed(10);
+  tab.minColWidths[2] = fixed(10);
+  print(tab);
+}