+++ example/table/styles.dart
@@ -3,8 +3,7 @@ import 'package:text_table/text_table.dart';
 main() {
   {
     print('Default style:');
-    final Table tab = table(['Player', 'Team', 'Goals'],
-        columnWidths: [fixed(25), fixed(25), fixed(10)])
+    final Table tab = table(['Player', 'Team', 'Goals'])
       ..row(['Messi', 'Barcelona FC', 80])
       ..row(['Christiano Ronaldo', 'Real Madrid', 30])
       ..row(['Luiz Suarez', 'Barcelona FC', 50]);
@@ -15,7 +14,6 @@ main() {
     print('Default rounded:');
     final Table tab = table(
       ['Player', 'Team', 'Goals'],
-      columnWidths: [fixed(25), fixed(25), fixed(10)],
       border: Border.defRound,
     )
       ..row(['Messi', 'Barcelona FC', 80])
@@ -27,7 +25,7 @@ main() {
   {
     print('No borders:');
     final Table tab = table(['Player', 'Team', 'Goals'],
-        columnWidths: [fixed(25), fixed(25), fixed(10)], border: Border.compact)
+        globalPadding: pad(after: 5), border: Border.compact)
       ..row(['Messi', 'Barcelona FC', 80])
       ..row(['Christiano Ronaldo', 'Real Madrid', 30])
       ..row(['Luiz Suarez', 'Barcelona FC', 50]);
@@ -38,7 +36,6 @@ main() {
     print('Simple:');
     final Table tab = table(
       ['Player', 'Team', 'Goals'],
-      columnWidths: [fixed(25), fixed(25), fixed(10)],
       border: Border.simple,
     )
       ..row(['Messi', 'Barcelona FC', 80])
@@ -51,7 +48,6 @@ main() {
     print('Single line:');
     final Table tab = table(
       ['Player', 'Team', 'Goals'],
-      columnWidths: [fixed(25), fixed(25), fixed(10)],
       border: Border.singleLine,
     )
       ..row(['Messi', 'Barcelona FC', 80])
@@ -64,7 +60,6 @@ main() {
     print('Double lines:');
     final Table tab = table(
       ['Player', 'Team', 'Goals'],
-      columnWidths: [fixed(25), fixed(25), fixed(10)],
       border: Border.doubleLines,
     )
       ..row(['Messi', 'Barcelona FC', 80])
@@ -77,7 +72,6 @@ main() {
     print('Double lines2:');
     final Table tab = table(
       ['Player', 'Team', 'Goals'],
-      columnWidths: [fixed(25), fixed(25), fixed(10)],
       border: Border.doubleLines2,
     )
       ..row(['Messi', 'Barcelona FC', 80])
@@ -90,7 +84,6 @@ main() {
     print('Double lines3:');
     final Table tab = table(
       ['Player', 'Team', 'Goals'],
-      columnWidths: [fixed(25), fixed(25), fixed(10)],
       border: Border.doubleLines3,
     )
       ..row(['Messi', 'Barcelona FC', 80])