+++ lib/text_table.dart
@@ -3,6 +3,6 @@
 /// More dartdocs go here.
 library text_table;
 
-export 'src/text_table_base.dart';
+export 'src/core/core.dart';
 export 'src/style/style.dart';
 export 'src/dimension/dimension.dart';